API: pass config explicitly
Ditch the global state config voodoo as much as possible. Writable
global state is not your friend. Instead, pass a config object
explicitly (rename config -> cfg). This is still a good thing, as
opposed to passing a lot of kwds thru layers of functions.

Also, add config.getcfg() to force generating a copy of the default
config when using this (as much as things can be forced in Python).
Relies on the cfg=config.getcfg() pattern.

Remove any default kwds like blocksize=cfg.blocksize. Most of the code
is used as a library and only rarely interactively (only for quick
tests). Therefore, rm all interactive magic and always be explicit about
what is being passed.