BUG: separate empty file/dir fpr
Before we used hashsum('') to define and detect both empty files (size
0) and empty dirs (no files). Also, we re-used the file hash for the dir
if there was only one file in it. This lead to the bug that only dirs
with zero or exactly one single empty file would be registered as
dir:empty . Other dirs with # empty files > 1 would not.

That fact was also falsely reflected in our test data where we have only
dirs called "empty_dir" with one file, so the bug never showed in tests.
Will be fixed in a later commit.

Now that we use the file size as well when hashing files, we add
EMPTY_FILE_FPR = hashsum('0') and EMPTY_DIR_FPR = hashsum(''). The
former is the result of hashing such a file w/ hash_file(), the latter
is a definition.