API: allow only limit % blocksize == 0
Remove hash_file_limit_core() and adjust_blocksize(). Employ very
simple and fast-to-check rule:

  (1) if blocksize > limit, then set blocksize = limit
  (2) require limit % blocksize == 0

Since in practice we always use multiples of KiB (1024 B) for blocksize
and limit, there is no real practical downside. With that we get rid of
the overly convenient but slow adjust_blocksize() and the
hash_file_limit_core() hack.